+++ convoread/config.py
@@ -30,6 +30,7 @@ config = {
     'LIVE_URL': '/api/live.json',
     'GROUPS_URL': '/api/groups.json',
     'TOPICS_URL': '/api/groups/{0}/topics.json',
+    'TOPIC_MESSAGES_URL': '/api/topics/{0}/messages.json',
     'CREATE_MSG_URL': '/api/topics/{0}/messages/create.json',
     'ENCODING': 'UTF-8',
 }