+++ CHANGES.rst
@@ -6,7 +6,9 @@ Changes
 0.4, 2011-02-12
 ---------------
 
-- Commands for selecting topics and posting new messages
+- Post new messages
+- Command ``/t`` for selecting topics
+- Command ``/ls`` for listing recent messages
 - Improved message formatting
 - Avatars in notifications
 - Various bug fixes